+++ src/Pages/Dashboard/Components/TxHistory/Components/index.ts
@@ -1,1 +1,2 @@
-export { default as TxChart } from './TxChart';
+export * from './TxChart';
+export * from './DownloadCsvModal';